requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP clients
httpx==0.26.0
requests==2.31.0

# Salesforce integration
simple-salesforce==1.12.5

# AI APIs
openai==1.10.0
anthropic==0.40.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and monitoring
structlog==24.1.0

# Async support
asyncio==3.4.3
aiohttp==3.9.1

# Browser automation
playwright==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # for testing FastAPI endpoints

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
